Modularize code
- Create a new module called simulator
- Create a submodule for constants and another one for resources.
