--- playerlist.py
@@ -1,3 +1,0 @@
-import pandas as pd
-
-df = pd.read_pickle("playerdata")
