+++ simulator/__main__.py
@@ -1,0 +1,4 @@
+from simulator import app
+
+if __name__ == '__main__':
+    app.run()
