+++ simulator/constants/managers.py
