+++ simulator/constants/odds.py
