+++ simulator/resources/player_data
